Fix compiler warning.
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@811 2e953b5c-c64d-0410-be54-f773e93e544c

+++ smartgame/SgUctTree.cpp
@@ -304,7 +304,7 @@ void SgUctTree::MergeChildren(std::size_t allocatorId, const SgUctNode& node,
     // run past the end of a node's children, which can happen if one
     // is created between the two statements below. We modify node in
     // such a way so as to avoid that.
-    if (nonConstNode.NuChildren() < nuNewChildren)
+    if (nonConstNode.NuChildren() < (int)nuNewChildren)
     {
         nonConstNode.SetFirstChild(newFirstChild);
         nonConstNode.SetNuChildren(nuNewChildren);
